Data on maternal smoking and infant health

Data on maternal smoking and infant health

babies

Format

A data frame of 1236 rows of individual mothers.

id

Identification number

pluralty

Marked 5 for single fetus, otherwise number of fetuses

outcome

Marked 1 for live birth that survived at least 28 days

date

Birth date where 1096 is January 1st, 1961

birthday

Birth date in mm-dd-yyyy format

gestation

Length of gestation in days, marked 999 if unknown

sex

Infant's sex, where 1 is male, 2 is female, and 9 is unknown

wt

Birth weight in ounces, marked 999 if unknown

parity

Total number of previous pregnancies including fetal deaths and stillbirths, marked 99 if unknown

race

Mother's race where 0-5 is white, 6 is Mexican, 7 is Black, 8 is Asian, 9 is mixed, and 99 is unknown

age

Mother's age in years at termination of pregnancy, 99=unknown

ed

Mother's education 0= less than 8th grade, 1 = 8th -12th grade - did not graduate, 2= HS graduate--no other schooling , 3= HS+trade, 4=HS+some college 5= College graduate, 6&7 Trade school HS unclear, 9=unknown

ht

Mother's height in inches to the last completed inch, 99=unknown

wt_1

Mother prepregnancy wt in pounds, 999=unknown

drace

Father's race, coding same as mother's race

dage

Father's age, coding same as mother's age

ded

Father's education, coding same as mother's education

dht

Father's height, coding same as for mother's height

dwt

Father's weight coding same as for mother's weight

marital

0= legally separated, 1=married, 2= divorced, 3=widowed, 5=never married

inc

Family yearly income in $2500 increments 0 = under 2500, 1=2500-4999, ..., 8= 12,500-14,999, 9=15000+, 98=unknown, 99=not asked

smoke

Does mother smoke? 0=never, 1= smokes now, 2=until current pregnancy, 3=once did, not now, 9=unknown

time

If mother quit, how long ago? 0=never smoked, 1=still smokes, 2=during current preg, 3=within 1 yr, 4= 1 to 2 years ago, 5= 2 to 3 yr ago, 6= 3 to 4 yrs ago, 7=5 to 9yrs ago, 8=10+yrs ago, 9=quit and don't know, 98=unknown, 99=not asked

number

Number of cigs smoked per day for past and current smokers 0=never, 1=1-4, 2=5-9, 3=10-14, 4=15-19, 5=20-29, 6=30-39, 7=40-60, 8=60+, 9=smoke but don't know, 98=unknown, 99=not asked

Source

Data on maternal smoking and infant health from https://www.stat.berkeley.edu/~statlabs/labs.html