Data from the Coffee Quality Institute's review pages in January 2018

1,340 digitized reviews on coffee samples from https://database.coffeeinstitute.org/.

coffee_ratings

Format

A data frame of 1,340 rows representing each sample of coffee.

total_cup_points

Number of points in final rating (scale of 0-100)

species

Species of coffee bean plant (Arabica or Robusta)

owner

Owner of coffee plant farm

country_of_origin

Coffee bean's country of origin

farm_name

Name of coffee plant farm

lot_number

Lot number for tested coffee beans

mill

Name of coffee bean's processing facility

ico_number

International Coffee Organization number

company

Name of coffee bean's company

altitude

Altitude at which coffee plants were grown

region

Region where coffee plants were grown

producer

Name of coffee bean roaster

number_of_bags

Number of tested bags

bag_weight

Tested bag weight

in_country_partner

Partner for the country

harvest_year

Year the coffee beans were harvested

grading_date

Day the coffee beans were graded

owner_1

Owner of the coffee beans

variety

Variety of the coffee beans

processing_method

Method used for processing the coffee beans

aroma

Coffee aroma rating

flavor

Coffee flavor rating

aftertaste

Coffee aftertaste rating

acidity

Coffee acidity rating

body

Coffee body rating

balance

Coffee balance rating

uniformity

Coffee uniformity rating

clean_cup

Cup cleanliness rating

sweetness

Coffee sweetness rating

cupper_points

Cupper Points, an overall rating for the coffee

moisture

Coffee moisture content

category_one_defects

Number of category one defects for the coffee beans

quakers

Number of coffee beans that don't dark brown when roasted

color

Color of the coffee beans

category_two_defects

Number of category two defects for the coffee beans

expiration

Expiration date of the coffee beans

certification_body

Entity/Institute that certified the coffee beans

certification_address

Body address of certification for coffee beans

certification_contact

Certification contact for coffee beans

unit_of_measurement

Unit of measurement for altitude

altitude_low_meters

Lower altitude level coffee beans grow at

altitude_high_meters

Higher altitude level coffee beans grow at

altitude_mean_meters

Average altitude level coffee beans grow at

Source

Coffee Quality Institute. Access cleaned data available at https://github.com/jldbc/coffee-quality-database