House Sales in King County, USA

This dataset contains house sale prices for King County, which includes Seattle. It includes homes sold between May 2014 and May 2015. This dataset was obtained from Kaggle.com https://www.kaggle.com/harlfoxem/housesalesprediction/data

Usage

house_prices

Format

A data frame with 21613 observations on the following 21 variables.

id

a notation for a house

date

Date house was sold

price

Price is prediction target

bedrooms

Number of Bedrooms/House

bathrooms

Number of bathrooms/bedrooms

sqft_living

square footage of the home

sqft_lot

square footage of the lot

floors

Total floors (levels) in house

waterfront

House which has a view to a waterfront

view

Has been viewed

condition

How good the condition is (Overall)

grade

overall grade given to the housing unit, based on King County grading system

sqft_above

square footage of house apart from basement

sqft_basement

square footage of the basement

yr_built

Built Year

yr_renovated

Year when house was renovated

zipcode

zip code

lat

Latitude coordinate

long

Longitude coordinate

sqft_living15

Living room area in 2015 (implies– some renovations) This might or might not have affected the lotsize area

sqft_lot15

lotSize area in 2015 (implies– some renovations)

Source

Kaggle https://www.kaggle.com/harlfoxem/housesalesprediction. Note data is released under a CC0: Public Domain license.