International Power Lifting Results A subset of international powerlifting results.

International Power Lifting Results A subset of international powerlifting results.

ipf_lifts

Format

A data frame with 41,152 entries, one entry for individual lifter

name

Individual lifter name

sex

Binary sex (M/F)

event

The type of competition that the lifter entered

equipment

The equipment category under which the lifts were performed

age

The age of the lifter on the start date of the meet

age_class

The age class in which the filter falls

division

division of competition

bodyweight_kg

The recorded bodyweight of the lifter at the time of competition, to two decimal places

weight_class_kg

The weight class in which the lifter competed, to two decimal places

best3squat_kg

Maximum of the first three successful attempts for the lift

best3bench_kg

Maximum of the first three successful attempts for the lift

best3deadlift_kg

Maximum of the first three successful attempts for the lift

place

The recorded place of the lifter in the given division at the end of the meet

date

Date of the event

federation

The federation that hosted the meet

meet_name

The name of the meet

Source

This data is a subset of the open dataset Open Powerlifting