moderndive - Accompaniment Package to ModernDive: An Introduction to Statistical and Data Sciences via R

An accompaniment R package to ModernDive: An Introduction to Statistical and Data Sciences via R available at http://moderndive.com/, in particular wrapper functions targeted at novices to easily generate tidy linear regression output in Chapter 6: Data Modeling with Regression.

Examples

# Example usage:
library(moderndive)
library(dplyr)

# Regression models
mpg_model <- lm(mpg ~ hp, data = mtcars)
mpg_mlr_model <- lm(mpg ~ hp + wt, data = mtcars)
mpg_mlr_model2 <- lm(mpg ~ hp + cyl, data = mtcars)

# Regression tables
get_regression_table(model = mpg_model)

#> # A tibble: 2 x 7
#>   term      estimate std_error statistic p_value conf_low conf_high
#>   <chr>        <dbl>     <dbl>     <dbl>   <dbl>    <dbl>     <dbl>
#> 1 intercept   30.1        1.63     18.4        0   26.8      33.4  
#> 2 hp          -0.068      0.01     -6.74       0   -0.089    -0.048

get_regression_table(mpg_mlr_model, digits = 4, print = TRUE)

#> 
#> 
#> |term      | estimate| std_error| statistic| p_value| conf_low| conf_high|
#> |:---------|--------:|---------:|---------:|-------:|--------:|---------:|
#> |intercept |  37.2273|    1.5988|   23.2847|  0.0000|  33.9574|   40.4972|
#> |hp        |  -0.0318|    0.0090|   -3.5187|  0.0015|  -0.0502|   -0.0133|
#> |wt        |  -3.8778|    0.6327|   -6.1287|  0.0000|  -5.1719|   -2.5837|

# Regression points. For residual analysis for example
get_regression_points(mpg_mlr_model2)

#> # A tibble: 32 x 6
#>       ID   mpg    hp   cyl mpg_hat residual
#>    <int> <dbl> <dbl> <dbl>   <dbl>    <dbl>
#>  1     1  21     110     6    21.2   -0.217
#>  2     2  21     110     6    21.2   -0.217
#>  3     3  22.8    93     4    26.1   -3.27 
#>  4     4  21.4   110     6    21.2    0.183
#>  5     5  18.7   175     8    15.4    3.26 
#>  6     6  18.1   105     6    21.3   -3.21 
#>  7     7  14.3   245     8    14.1    0.194
#>  8     8  24.4    62     4    26.7   -2.26 
#>  9     9  22.8    95     4    26.0   -3.23 
#> 10    10  19.2   123     6    21.0   -1.77 
#> # ... with 22 more rows

# Regression summaries
get_regression_summaries(mpg_model)

#> # A tibble: 1 x 8
#>   r_squared adj_r_squared   mse  rmse sigma statistic p_value    df
#>       <dbl>         <dbl> <dbl> <dbl> <dbl>     <dbl>   <dbl> <dbl>
#> 1     0.602         0.589  14.0  3.74  3.86      45.5       0     2

# Can also use `%>%`
mpg_model %>% get_regression_summaries(digits = 5, print = TRUE)

#> 
#> 
#> | r_squared| adj_r_squared|      mse|     rmse|   sigma| statistic| p_value| df|
#> |---------:|-------------:|--------:|--------:|-------:|---------:|-------:|--:|
#> |   0.60244|       0.58919| 13.98982| 3.740297| 3.86296|   45.4598|       0|  2|