One permutation/shuffle of promotions

Shuffled/permuted data from a 1970's study on whether gender influences hiring recommendations.

promotions_shuffled

Format

A data frame with 48 observations on the following 3 variables.

id

Identification variable used to distinguish rows.

gender

shuffled/permuted (binary) gender: a factor with two levels `male` and `female`

decision

a factor with two levels: `promoted` and `not`

See also

Examples

library(dplyr)
glimpse(promotions)

#> Rows: 48
#> Columns: 3
#> $ id       <int> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 1…
#> $ decision <fct> promoted, promoted, promoted, promoted, promoted, promoted, …
#> $ gender   <fct> male, male, male, male, male, male, male, male, male, male, …

glimpse(promotions_shuffled)

#> Rows: 48
#> Columns: 3
#> $ id       <int> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 1…
#> $ decision <fct> promoted, promoted, promoted, promoted, promoted, promoted, …
#> $ gender   <fct> female, female, male, female, male, male, male, female, male…