Teaching evaluations at the UT Austin

The data are gathered from end of semester student evaluations for a large sample of professors from the University of Texas at Austin. In addition, six students rate the professors' physical appearance. The result is a data frame where each row contains a different course and each column has information on either the course or the professor https://www.openintro.org/stat/data/?data=evals

evals

Format

A data frame with 463 observations on the following 13 variables.

ID

Identification variable used to distinguish rows.

score

Average professor evaluation score: (1) very unsatisfactory - (5) excellent.

age

Age of professor.

bty_avg

Average beauty rating of professor.

gender

Gender of professor: female, male.

ethnicity

Ethnicity of professor: not minority, minority.

language

Language of school where professor received education: English or non-English.

rank

Rank of professor: teaching, tenure track, tenured.

pic_outfit

Outfit of professor in picture: not formal, formal.

pic_color

Color of professor’s picture: color, black & white.

cls_did_eval

Number of students in class who completed evaluation.

cls_students

Total number of students in class.

cls_level

Class level: lower, upper.

Source

Çetinkaya-Rundel M, Morgan KL, Stangl D. 2013. Looking Good on Course Evaluations. CHANCE 26(2). http://chance.amstat.org/2013/04/looking-good/

Examples

library(dplyr)
glimpse(evals)

#> Observations: 463
#> Variables: 13
#> $ ID           <int> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16,...
#> $ score        <dbl> 4.7, 4.1, 3.9, 4.8, 4.6, 4.3, 2.8, 4.1, 3.4, 4.5, 3.8,...
#> $ age          <int> 36, 36, 36, 36, 59, 59, 59, 51, 51, 40, 40, 40, 40, 40...
#> $ bty_avg      <dbl> 5.000, 5.000, 5.000, 5.000, 3.000, 3.000, 3.000, 3.333...
#> $ gender       <fct> female, female, female, female, male, male, male, male...
#> $ ethnicity    <fct> minority, minority, minority, minority, not minority, ...
#> $ language     <fct> english, english, english, english, english, english, ...
#> $ rank         <fct> tenure track, tenure track, tenure track, tenure track...
#> $ pic_outfit   <fct> not formal, not formal, not formal, not formal, not fo...
#> $ pic_color    <fct> color, color, color, color, color, color, color, color...
#> $ cls_did_eval <int> 24, 86, 76, 77, 17, 35, 39, 55, 111, 40, 24, 24, 17, 1...
#> $ cls_students <int> 43, 125, 125, 123, 20, 40, 44, 55, 195, 46, 27, 25, 20...
#> $ cls_level    <fct> upper, upper, upper, upper, upper, upper, upper, upper...