Get regression points

Output information on each point/observation used in an lm() regression in "tidy" format. This function is a wrapper function for broom::augment() and renames the variables to have more intuitive names.

get_regression_points(model, digits = 3, print = FALSE, newdata = NULL)

Arguments

model
an lm() model object

digits
number of digits precision in output table

print
If TRUE, return in print format suitable for R Markdown

newdata
A new data frame of points/observations to apply model to obtain new fitted values and/or predicted values y-hat. Note the format of newdata must match the format of the original data used to fit model.

Value

A tibble-formatted regression table of outcome/response variable, all explanatory/predictor variables, the fitted/predicted value, and residual.

See also

Examples

library(moderndive)
library(dplyr)
library(tibble)

# Fit lm() regression:
mpg_model <- lm(mpg ~ cyl, data = mtcars)

# Get information on all points in regression:
get_regression_points(mpg_model)

#> # A tibble: 32 x 5
#>       ID   mpg   cyl mpg_hat residual
#>    <int> <dbl> <dbl>   <dbl>    <dbl>
#>  1     1  21       6    20.6    0.37 
#>  2     2  21       6    20.6    0.37 
#>  3     3  22.8     4    26.4   -3.58 
#>  4     4  21.4     6    20.6    0.77 
#>  5     5  18.7     8    14.9    3.82 
#>  6     6  18.1     6    20.6   -2.53 
#>  7     7  14.3     8    14.9   -0.578
#>  8     8  24.4     4    26.4   -1.98 
#>  9     9  22.8     4    26.4   -3.58 
#> 10    10  19.2     6    20.6   -1.43 
#> # ... with 22 more rows

# Create training and test set based on mtcars: 
mtcars <- mtcars %>%
  rownames_to_column(var = "model")
training_set <- mtcars %>%
  sample_frac(0.5)
test_set <- mtcars %>%
  anti_join(training_set, by = "model")

# Fit model to training set:
mpg_model_train <- lm(mpg ~ cyl, data = training_set)

# Make predictions on test set:
get_regression_points(mpg_model_train, newdata = test_set)

#> # A tibble: 16 x 5
#>       ID   mpg   cyl mpg_hat residual
#>    <int> <dbl> <dbl>   <dbl>    <dbl>
#>  1     1  21       6    20.8    0.187
#>  2     2  24.4     4    26.2   -1.75 
#>  3     3  17.8     6    20.8   -3.01 
#>  4     4  16.4     8    15.5    0.929
#>  5     5  15.2     8    15.5   -0.271
#>  6     6  10.4     8    15.5   -5.07 
#>  7     7  10.4     8    15.5   -5.07 
#>  8     8  14.7     8    15.5   -0.771
#>  9     9  32.4     4    26.2    6.25 
#> 10    10  30.4     4    26.2    4.25 
#> 11    11  15.5     8    15.5    0.029
#> 12    12  19.2     8    15.5    3.73 
#> 13    13  26       4    26.2   -0.154
#> 14    14  19.7     6    20.8   -1.11 
#> 15    15  15       8    15.5   -0.471
#> 16    16  21.4     4    26.2   -4.75