Get regression points
Source: R/regression_functions.R

Wrapper for tidy lm() regression function fit output

get_regression_points(model, digits = 3, print = FALSE, newdata = NULL)

Arguments

model
a model object (Currently only lm is supported)

digits
number of digits precision in output table

print
If TRUE, return in print format suitable for R Markdown

newdata
a new data frame to fit the regression to/make predictions on.

Value

A tibble or nicely formatted table

Examples

library(dplyr)
library(tibble)

# Get information on all points in regression
mpg_model <- lm(mpg ~ cyl, data = mtcars)
get_regression_points(mpg_model)

#> # A tibble: 32 x 5
#>       ID   mpg   cyl mpg_hat residual
#>    <int> <dbl> <dbl>   <dbl>    <dbl>
#>  1     1  21.0    6.    20.6    0.370
#>  2     2  21.0    6.    20.6    0.370
#>  3     3  22.8    4.    26.4   -3.58 
#>  4     4  21.4    6.    20.6    0.770
#>  5     5  18.7    8.    14.9    3.82 
#>  6     6  18.1    6.    20.6   -2.53 
#>  7     7  14.3    8.    14.9   -0.578
#>  8     8  24.4    4.    26.4   -1.98 
#>  9     9  22.8    4.    26.4   -3.58 
#> 10    10  19.2    6.    20.6   -1.43 
#> # ... with 22 more rows

# Convert rownames to column
data("mtcars")
mtcars <- mtcars %>%
  rownames_to_column(var = "model")

# Create training and test set.
training_set <- mtcars %>%
  sample_frac(0.5)
test_set <- mtcars %>%
  anti_join(training_set, by = "model")

# Fit model to training set, predict on test set
mpg_model_train <- lm(mpg ~ cyl, data = training_set)
get_regression_points(mpg_model_train, newdata = test_set)

#> # A tibble: 16 x 3
#>      mpg   cyl mpg_hat
#>    <dbl> <dbl>   <dbl>
#>  1  21.0    6.    20.8
#>  2  24.4    4.    26.2
#>  3  17.8    6.    20.8
#>  4  16.4    8.    15.5
#>  5  15.2    8.    15.5
#>  6  10.4    8.    15.5
#>  7  10.4    8.    15.5
#>  8  14.7    8.    15.5
#>  9  32.4    4.    26.2
#> 10  30.4    4.    26.2
#> 11  15.5    8.    15.5
#> 12  19.2    8.    15.5
#> 13  26.0    4.    26.2
#> 14  19.7    6.    20.8
#> 15  15.0    8.    15.5
#> 16  21.4    4.    26.2